test: move job stream endpoint test cases to the correct IT class
